Add benchmark instrumentation, struct packing, and deterministic seeding
Add per-move timing and node counting around genmove calls, plus a
COMMAND_BENCHMARK handler that self-plays N moves at a specified level
for repeatable A/B performance testing.

Pack worm_data, dragon_data, eye_data, half_eye_data, and move_data
structs using uint8_t/uint16_t/int16_t for fields with small value
ranges, saving ~123KB of PSRAM. Fix compare_move_reasons() to cast
through int16_t* after narrowing move_data.reason[] from int to int16_t.

Remove CONFIG_USE_ESP_RANDOM Kconfig option. Normal gameplay seeds the
TGFSR PRNG via esp_random() on ESP32; benchmarks force seed 1 for
determinism. The TGFSR generator is now always used for random number
generation.

+++ engine/move_reasons.h
@@ -84,6 +84,8 @@
 
 #define MAX_TRACE_LENGTH  160
 
+#include <stdint.h>
+
 #define HUGE_MOVE_VALUE 10.0*MAX_BOARD*MAX_BOARD
 
 struct move_reason {
@@ -120,7 +122,7 @@ struct move_data {
   float max_territory;        /* Maximum territorial value. */
   float randomness_scaling;   /* Increase to randomize this move. */
 
-  int reason[MAX_REASONS]; /* List of reasons for a move. */
+  int16_t reason[MAX_REASONS]; /* List of reasons for a move (-1 = unused). */
   int move_safety;         /* Whether the move seems safe. */
   int worthwhile_threat;   /* Play this move as a pure threat. */
   float random_number;     /* Random number connected to this move. */
